move inline functions to utils.js
